refactor(a2-480): remove log, move formatComments to correct dir
