+++ packages/forms-web-app/src/controllers/selected-appeal/ip-comment-details/index.js
@@ -5,7 +5,7 @@ const { apiClient } = require('../../../lib/appeals-api-client');
 const { formatTitleSuffix } = require('../../../lib/selected-appeal-page-setup');
 const { determineUser } = require('../../../lib/determine-user');
 const { getUserFromSession } = require('../../../services/user.service');
-const { formatComments } = require('../final-comments-details/sort-comment-details');
+const { formatComments } = require('./sort-comment-details');
 
 /**
  * Shared controller for /manage-appeals/:caseRef/appellant-final-comments, manage-appeals/:caseRef/final-comments
@@ -42,7 +42,6 @@ exports.get = (layoutTemplate = 'layouts/no-banner-link/main.njk') => {
 
 		const comments = await req.appealsApiClient.getInterestedPartyComments(appealNumber);
 		const formattedComments = formatComments(comments);
-		console.log('###', formattedComments);
 
 		const headlineData = formatHeadlineData(caseData, userType);
